+++ client/src/context/Provider.js
@@ -1,65 +1,77 @@
-// src/context/Provider.js
-import React, { createContext, useState, useContext } from 'react';
+import React, { createContext, useState, useContext, useEffect } from 'react';
 
+// Create the context
 const AppContext = createContext();
 
+// Context Provider Component
 export const Provider = ({ children }) => {
-  const [user, setUser] = useState(null);
-  const [error, setError] = useState(null);
+    const [user, setUser] = useState(null); // holds the current logged-in user
+    const [error, setError] = useState(null); // optional for error tracking
+    const [fundi, setFundi] = useState()
+    const [services, setServices] = useState([]);
 
-//   // Other global state placeholders
-//   const [services, setServices] = useState([]);
-//   const [orders, setOrders] = useState([]);
+    const retrieve_services = async () => {
+        try {
+        const res = await fetch("/services");
+        const data = await res.json();
 
-//   // Signup function
-//   const signupUser = async ({ username, email, password, phone_number }) => {
-//     try {
-//       const res = await fetch('/signup', {
-//         method: 'POST',
-//         headers: { 'Content-Type': 'application/json' },
-//         credentials: 'include',
-//         body: JSON.stringify({ username, email, password, phone_number }),
-//       });
+        if (res.ok) {
+            setServices(data);
+        } else {
+            setError(data.error || "Failed to retrieve services.");
+        }
+        } catch (err) {
+        console.error("Service-retrieval error:", err);
+        setError("Server error.");
+        }
+    };
 
-//       const data = await res.json();
+    // Fetch fundi details
+    const fetchFundi = async (id) => {
+    if (!id) {
+        alert('Provider called without valid fundi id');
+        return;
+    }
+    try {
+        const res = await fetch(`/fundis/${id}`);
+        const data = await res.json();
 
-//       if (res.ok) {
-//         setUser(data);
-//         return { success: true, data };
-//       } else {
-//         setError(data.error || 'Signup failed.');
-//         return { success: false, error: data.error };
-//       }
-//     } catch (err) {
-//       console.error('Signup error:', err);
-//       setError('Server error.');
-//       return { success: false, error: 'Server error.' };
-//     }
-//   };
+        if (res.ok) {
+        setFundi(data);
+        } else {
+        setError(data.error || "Failed to fetch fundi");
+        console.error("Failed to fetch fundi:", data);
+        }
+    } catch (err) {
+        // console.error("Error fetching fundi:", err);
+        console.log('error fetching fundi')
+        setError("Server error while fetching fundi");
+    }
+    };
 
-  return (
+    return (
     <AppContext.Provider
-      value={{
+        value={{
         user,
         setUser,
-        // error,
-        // setError,
-        // signupUser,
-        // services,
-        // setServices,
-        // orders,
-        // setOrders,
-        // You can keep adding shared state/methods here
-      }}
+        fetchFundi,
+        retrieve_services,
+        services,
+        fundi,
+        error,
+        setError,
+        }}
     >
-      {children}
+        {children}
     </AppContext.Provider>
-  );
+    );
 };
 
-// Custom hook to access context
+// Custom hook to use the context safely
 export const useTheContext = () => {
   const context = useContext(AppContext);
-  if (!context) throw new Error('useTheContext must be used within a Provider');
+  if (!context) {
+    throw new Error('useTheContext must be used within a Provider');
+  }
   return context;
 };
